+++ src/app/(route)/hello/_components/FEATURES_CONST.tsx
@@ -1,4 +1,4 @@
-import { FourthSection, HeroSection, SecondSection, ThirdSection } from "../_components/ImageSlots";
+import { FourthSection, HeroSection, SecondSection, ThirdSection } from "./ImageSlots";
 
 export const FEATURES = [
   {
